fix(avalonia): address Copilot review on AI edit/write (#760)
- UpdateRow: preserve the row's per-offset comment (DisAseemble(bytes,0) was
  pulling Comment from CommentCache.At(0) — wrong offset).
- WriteScript: use rom.write_range for a single undo entry instead of
  per-byte write_u8 (avoids bloating the undo buffer).
- Write_Click: guard an empty/never-loaded model before opening the undo
  scope (no ghost undo entry / misleading "length changed" message); add
  AIScriptViewModel.HasDisassembly.
- Test: cast byte[] length to (int) in AiDisasmEnv.RomSlice for clarity.

Goal: resolve all Avalonia-vs-WinForms GUI parity gaps.

+++ FEBuilderGBA.Avalonia.Tests/AIScriptDisassemblyTests.cs
@@ -351,7 +351,7 @@ namespace FEBuilderGBA.Avalonia.Tests
         /// before/after byte comparisons in the #760 write tests.</summary>
         public byte[] RomSlice(uint addr, uint length)
         {
-            var slice = new byte[length];
+            var slice = new byte[(int)length];
             Array.Copy(_rom.Data, (int)addr, slice, 0, (int)length);
             return slice;
         }
